Fix Escape resetting EditIcon drag sending "0".
This better matches the expected situations where this happens.

Basically if you are editing an icon with no offset, then drag it and escape key out it sets the offsets back to 0, 0 instead of both being blank, thus when closing the window it prompts a close without saving. Also it is just visually weird to have it update on escaping.