Fixed offsets not accepting undefined.
Found a bug where having existing offsets and then updating from dragging to undefined would not update the offsets in the edit icon menu, but would update the actual offsets, and appear correctly.

Basically user could edit an icon with no offsets, drag it and drop it, then drag it and escape to reset to where it was when edit icon was opened (both blank inputs). But it would keep the values from the last time they moved it, even though it correctly places it and would correctly save it. (purely visual bug in the inputs).