Fix EditIcon shift drag snapping for Icons with default sizes.
Fixed a problem where icons which did not have a custom width/height set would fallback to 64x64 for snapping.
Thus if the setting for default icon size in settings was say 200. It would snap at the 64x64 for the bottom/right edges instead of the actual bottom/right of the icon leading it being freely placed when expected to snap, and randomly snapping at 30%.