ContextMenu clicks on Icon boxes, now show mostly Desktop options.
Changed the ContextMenu when clicking on an iconBox to match a desktop right click instead. As it is assumed that if you wanted the icon options you would click on the icon itself rather than just the icon box.

Still keeps the "Edit ICON_NAME" at the top to help signify that this iconBox is taken.
But the rest of the contextMenu matches the Desktop contextMenu with options like changing backgrounds, settings, hide options, debugging, reload etc.

This also helps users who have large icons and a mostly filled desktop of icons as in the past they would have to find an iconBox which has no icon or they could lock themselves out of things like the debug options, reloading desktop, and settings/change background (if they were unaware of the header buttons).