Fix EnsureUniqueID not being case-insensitive.
Fixed a scenario where naming an icon "Icon" when you have another icon named "icon" (only difference is lowercase letter), would return that the name "Icon" is okay because it did not find another "Icon" (it was case sensitive).
The problem with this is that the folders are not case sensitive, so you cannot have an "icon" and an "Icon" folder. Thus when trying to copy the folder to "Icon" it would fail, thus breaking the path for the icon image, and leaving the old id folder.

Now it becomes "Icon_1" and ensures that it keeps its folder up to date.