Fix new background keeping selected background previewed.
Fixed a problem with backgroundPath check which uses the out of date backgroundPath = "" thus on a new background with no id/path, it would always fail this check.

This problem stems from the useEffect and useState functionality where if we wanted to check backgroundPath we would need it as a dependency. However as this is a window.electron.on listener, if it was a dependency then every time it was changed it would add another listener, endlessly.

At worst this preview update only runs once every time the Background File is changed. Which if changed it restarts from the beginning anyways. And for the most part it preview-background-update only calls with id set if it is changing. As things like volume call preview without an id.