EditBackground "Escape" to close.
Pressing "Escape" with EditBackground in focus now attempts closing the window.
This uses the base handleClose which checks for changes and will ask for confirmation on closing when changes from the inital background exist.

Also re-added handleClose dependency to BackgroundSelect, as handleClose is a async function, and if it is expanded later on, not having the dependency could cause issues. (For instance any value that changes after mount without the dependency instead uses the value it had on mount, which is why the "save without changes" windows require it to detect changes).