"Escape" key press will close smallWindow.
"Escape" key can now close a smallWindow.
This sends back an "Escape" response instead of a button text.
This does require smallWindow responses to properly support an escape as a basic response. This works for most, but there might be some smallWindows which are set up to require a button response. Will look into these if they exist.