smallWindow "closable" argument.
Can now create a smallWindow with a boolean "closeable" which determines if it is allowed to close via "Escape" or shows the X button to close.

Essentially call with false when you want the user to specifically pick one button response. (Logic should still be added for a default response just in case, but this more effectively guides the user into making a selection).

Also gave Settings a pendingSave state to stop it from being closable while pending the save. (Could previously click into Settings window then escape/X close -> confirm close without saving to avoid choosing an option for import as new or import as saved date smallWindow.

I am aware of the ability to close via header right click -> close, which also has the problem of being able to minimize subwindows. Both of which cause problems like destroying a reference to subWindows not allowing them to be opened again, etc. Will look into that soon. As of now doing this just requires a restart.