Disabling windows context-menu in SubWindow/SmallWindow.
Weird quirk of electron -webkit-app-region: drag; is that it sets anything with it to a "title-bar" which Windows then attaches a default window contextMenu into which is not natively supported with electron.
This contextMenu has things like "Move", "Close", "Minimize", "Maximize" and a few more.
Particularly the "Close" and "Minimize" are dangerous because "Close" destroys the object and causes errors in subWindows; and "Minimize" tries to minimize it but as it is a subWindow of the main BrowserWindow it does not exactly know how to handle itself so it just appears as a small black rectangle in the bottom of the main BrowserWindow with no native way to re-open it.

For this reason its better to just hook into the windows menu and disable it with this .hookWindowMessage function.

Shout out to qjl1569 for this work-around:
https://github.com/electron/electron/issues/24893#issuecomment-1109262719

As I'm not planning on using the contextMenu natively, I removed the menu and now right clicking just does nothing in the webkit drag area.

The main browserWindow is fine, as these are natively supported, so "close" calls the native mainWindow.on("hide"...) and same for minimize which is the same as the actual header buttons.