Fix Tag arrays causing getChanges to return true even when no changes.
When attempting to close it is possible that the current array of tags is in a different order than the original summary of tags, even if no changes were made.

To solve for this it now sorts both arrays then checks them to ensure that all values match as expected.