Testing phase importIconsFromDesktop.
For now it gets the first file/folder it finds on the desktop, and attempts to import it to 0,0. THIS WILL OVERWRITE.

Still needs a lot of work. This is very early in development which is why it can still only be ran via ipc call (await window.electron.importIconsFromDesktop();).

For instance:
Only imports first file/folder found.
Only places at 0,0.
Images use generateIcon which gets the default "image" windows icon instead of just using itself. (Think it needs saveToBackgroundIDFile like implementation).
Will use visibleRows/Cols to place icons on the grid eventually (while avoiding the existing icons).
Need more logic to either allow the user to select the icons if the file generates multiple icons, or a better approach which selects the most likely wanted icon.