Fix unknown mimetype failing to add icon.
Fixed a mimeType return of unknown from just returning to instead just continue with "unknown" mimeType. The only real check we are doing is just is it an image anyways.

In the case that it is unknown then it just generates the icon like normal, of which it would just generate the default windows icon for a file/filetype.