Only imports icons if they do not exist on current profile desktop.
importIconsFromDesktop now imports the first icon from the desktop which is not referenced by any current DesktopIcons. It checks this by the programLink.
So if you say already have an icon with the same name but different path it will still import it.
But if you have 2 icons with the same file reference it will not. (I think this might be the default behavior I think it makes general sense, trying to think of scenarios where some other setup might be more ideal... might swap it over to both icon name and program path.)

So technically as long as you move the icons before the next ipc call you can import your entire desktop...