importIconsFromDesktop nowmust match Name and Program path.
Previously it was only checking programLink path if it also existed in some other icon. This has been made more strict to also check that the icon is named the same as the name it would be imported by.

For instance a Desktop file like: "someFile.lnk" with a filepath ".../desktop/someFile.lnk" imports as long as there is no icon named: "someFile.lnk" which also has the filePath ".../desktop/someFile.lnk".

So if a DesktopIcon exists: "someFile.lnk" but the path is ".../someOtherFile.lnk" it will still import it.
Conversely if an icon exists with ".../desktop/someFile.lnk" program path, but the name is "someOtherFileName.lnk" it will still import it.

For it to be detected as already being a DesktopIcon it must fit both these categories.

In the future when Icons Profile is fully implemented there will be ways to see which icons Directly exist, which icons seem to be modified slightly, and which are completely different to any icons on the Desktop (or when comparing two profiles eventually).

Note:
Importing from desktop currently does not look for name conflicts for assigning id, thus it is possible to overwrite/merge two icons. So avoid importing ones with the same name as a DesktopIcon for now.