ensure newly imported icons have unique IDs.
Fixed a problem mentioned yesterday where it is possible to import icons with the same ID as existing icons leading to them being "synched" and extremely buggy due to being synched.
It now applies the same function to ensure the icon has a unique ID as it does when creating/saving icons. (ensuring that it creates and saves as a unique icon).