Fix for 2 identical profiles, but with different images not previewing.
Fixed a problem with how the preview updating works where it only reloads the icons if the iconsById and posMap get updated.

This can easily be found by creating a new profile and copying from the old profile, then delete/replace an icons image and it would previously not reload the icons, instead leaving them as they are despite a different image for an icon.

This is a very 0.01% bug, but directly happened to me, so I added a quick check.

For reference due to the setup with iconsById being id -> DesktopIcon vs posIndex which is "row,col" -> id, checking posIndex is much faster (since it does not require also comparing DesktopIcon objects) and covers more bases than checking iconsById.