Prevent accidental importing of Desktop into non-current profile.
If getRendererState for some reason fails it does not continue with importing the desktop files, and instead shows a smallWindow error.

This prevents it from accidentally importing into the "default" profile (which would happen when getRendererState failed and it fell into the fallback to "".

This is only a theoretical change. Before this commit the only way I could get this to trigger was repeated setRendererStates profile to "" via direct ipc call.
But it is now impossible to set profile to "" as it instead now defaults to "default" instead, so the only way this triggers is some major error where getRendererState just fails entirely.