export getDesktopUniqueFiles() and simplify logic till after confirmation.
Exported the check which determines which Desktop files are not already Desktop Icons so we can call it later to tell which files would be unique if imported (important for Desktop Icons Profile window later).

Also moved logic which only mattered on importing desktop files (takenCoords, maxRows/maxCols fetching) to AFTER confirmation for import. As it is not needed before confirmation and would simply be wasting compute if the user declined importing all.