Prevent "default" Desktop profile flicker on launch (for non-default profiles).
Now before the renderer shows it calls to update the rendererState with the saved backgrounds profile.

This means that the rendererState for profile is already set correctly before the program shows.

This fixes a problem where a non-default profile would still have the "default" profile appear for a few milliseconds before swapping to the saved profile.