Profile swapping IconsProfile.
Added a dropdown to swap profiles in IconsProfile.
This currently updates and saves the background bg.json file immediately on swapping profile. So when closing IconsProfile the currently previewed profile persists.
Also this makes it such that the total unique DesktopFiles found and the list of found DesktopFiles match up with the current profile.