SafeImage now falls back to current profile if not called with profile.
SafeImage now fetches the current rendererState profile if not called with one initially.

This fixes a bug in SelectIconWindow which caused no icons to ever properly display due to setting the path:
AppData\Roaming\AltDesktop\profiles\undefined\icons\...
Where it needs the correct profile, so it was always resolving to the fallback ? icon.

Not 100% satisfied with this, might go through and just make the actual call also pass the profile to avoid this awkward fallback which requires another ipc call PER icon.