Closing SelectIconWindow now reloads icon.
When highlighting(selecting) icons in SelectIconWindow it previews on the desktop, however if you closed out of the SelectIconWindow then it would previously not un-render the preview. So it would get stuck with the last highlighted icon.

This now re-renders it based on the current icon (combination of all fields object). Thus un-doing the external preview and correctly showing the icon it would show if saved.