FIX: openFileDialog button not returning .lnk paths.
Major annoyance for several months, finally found a way to fix it.

Basically the problem was the default dialog.showOpenDialog() with extensions = "*" (wildcard), auto resolves .lnk paths.

This is good in most cases but in the use case in this program its important to get the .lnk path instead as certain programs will bake command line arguments into the .lnk shortcut, which some programs require in order to work properly or at all (looking at you discord).

The absolute dumbest solution to the most annoying problem. Since "*" resolves .lnks, just add every extension alphanumerically....

So a problem caused by a wildcard can be fixed by adding a bunch of wildcards...

The functionality is 99% the same, the only difference is if you actually click the dropdown you see all 30+ extensions "a*", "b*", "c*", ... etc.
but unlikely to be noticed by 99% of users.