Removed "current" tab for Desktop Profile importing
Decided against adding it as it just confuses everything with placing a tab for editing the current profile among the tabs for importing icons. If this is re-implemented it will likely be a much different setup.
More likely to just rewrite the way desktop deletion works (allow selecting multiple icons via drag/ctrl click etc.)