getDesktopUniqueFiles also now returns alreadyImported.
Now returns 2 arrays, 1 being the filesToImport (unique files).
And the 2nd array being the array of already imported icons.

For now the only use of this is DesktopProfile just logging
also received data: ....
where it logs all the already imported icons. Later on this will display these as "already imported" with no import option. Also will eventually add a "imported but modified" which will show the files which are partially unique (either program path is different, or name is different).