Limit getDesktopUniqueFiles calls to only existing profile.
Instead of defaulting it to default and causing it to call getDesktopUniqueFiles, instead we can just default it to "" and it will set the profile from settings. Thus we can then only trigger the electron api call once when it properly has a profile. Previously this was calling ~3 times.
No direct noticeable impact, but should be a significant actual impact (removes 2 calls which browse both desktop directory, and fetch all icons of that profile, which it then compares.)