getDesktopUniqueFiles now returns partial matches.
Now returns two more arrays nameOnlyMatches, and pathOnlyMatches.

These are essentially "Modified" imported desktop files. I.e. where either it matches by name, or by program path. So for instance if an icon only has a different name but is still the same link this will eventually be in a "Modified" section where it is likely that the user has modified either the name or the path on purpose. To avoid confusion where a user imports an icon and then changes the name slightly.

For now difference only appears in calling getDesktopUniqueFiles() as DesktopProfile does not yet utilize them.