Append "..." if iconPath is a substring.
It was technically possible for an icon to appear as a "partial match" with a fully green name, and a fully green path.

Now if the name/path is a substring of the Desktop File it appends "..." to signify that it cuts off and does not finish.

This is more common when something like a Desktop Icon with the name "Program" instead of "Program.exe" is set.