Support for .url files (steam game shortcuts).
Simply resolves the saved IconFile in the .url and fetches the .ico file for use.

Not sure of anything else that uses .url files (not sure if it is strictly standardized with naming), so unsure if this works for all .url files. But works for steam (the most likely .url files used by users).