DesktopFileCompare and ProfileIconCompare types.
Added 2 distinct types.
DesktopFileCompare, used by getDesktopUniqueFiles
ProfileIconCompare, used by compareProfiles (not fully implemented yet).

Reasoning for the split is that the filesToImport is different for desktop files vs profiles. For instance from desktop we just need the name and path and rebuild the rest with some default parameters.
However for Profile comparison we are fetching icons which are fully defined. So we need to fetch the icon as well if we want to import it basically as is.