Ignoring extensions when comparing.
Added standardizedIconName function which attempts to remove the extension name from the name of the file/icon name.

This allows something like "2.png" to get recognized as the same as an icon with a name of "2". I.e. if it seems like the user just deleted the extension from the icon name then we should not recognize the file name as a "new" file we should look at importing. Because its already there, just with a shortened name.