compareProfiles re-write.
Simplified most of the returns into just DesktopIcons, which we can then get the names/data etc. from.

More efficiently compares the icons and returns a "modified" array with: otherIcon (DesktopIcon) with an array of differences which we can display as modified in the UI (not implemented yet).

Moved DesktopIcon close to top of types.d.ts (was originally but kept adding the new types above it...)

reminder call the ipc via (since ui not implemented yet):
await window.electron.compareProfiles("profile1", "profile2");