Testing Other Profiles (compareProfiles) DesktopProfile.
Mostly copied the code for the Desktop Files tab to Other Profiles tab for now.

Since these are icons, we can technically build them in a renderer and display them more naturally. Or at least improve the way they currently display. But for now just trying to create a way to visibly check compareProfiles which is better than having to manually call and sort through the ipc return.

Modified isn't fully implemented for instance just lists the programLink of the other profile for that icon as green, then lists all "differences" without spaces as yellow.