Fix file path green/yellow rendering (DesktopProfile).
Modified "modified" to return an array with both icons, and the differences. Instead of just returning the otherIcon.

This allows us to have both file paths so we can compare them like we do with desktop files.

Still a lot of kinks to work out of this as its pretty complicated. Need to just sit down and simplify and work out what exactly I will need from everything...