openInExplorer fix for improper paths.
Updated the openInExplorer ipc function to check and verify paths before just attempting to open the filePath/parent fallback.

For instance an improper path like "test" would attempt to open the "." directory (and fail). It seems that even fs.existsSync() returns true even when called with an invalid path.

Also added some error smallWindow popups when it fails due to wrong parent, or not a valid path.