Fix white flashing in DesktopProfile on first launch.
Weird bug, for some reason 100+ms delay + "#00000000" (transparent black) flashes white, but only for DesktopProfile.
Good news is that this now launches faster (no delay). Bad news is this flashes the dark color for a few frames before it fully loads the frontend for DesktopProfile, but is still better than a literal flashbang.