Clear icons before swapping profiles.
Fixed a large amount of warnings (from redraw attempts for all icons) on swapping between profiles.

Should be slightly more efficient as it just clears them and does not attempt to redraw them due to async with setting the new icons.

It now effectively trashes all of the icons before setting the new profile. Which stops safeFileProtocol from trying to re-draw the icons with the new profiles paths.

Previously would spam:
[2025-12-06 23:25:42] [WARN] [safeFileProtocol.ts] File not found: C:\Users\...\profiles\new_profile\icons\WizTree64.png
for every single icon which had an image.