Loading theme from settings.
Now loads the theme from settings.json (defaults to "system").

Simplified the system/non "dark" or "light" into a function resolveThemeScheme which returns "dark" or "light" (light if light, otherwise shows/logs warning and defaults to dark).

Also when the user Reloads the desktop (Reload Desktop in desktop right-click, it now reloads the theme from this file ensuring it is up to date).