Fix default desktop (no background) DesktopProfile file index.
When no background is selected and Desktop Profile window is opened, it now properly loads the "default" background Desktop Files.

Previously loaded nothing as it was relying on a "profile" being set, which fails for the "" profile, and was looking for an actuall profile "" instad of the "default" profile.

I think I might need to go back and just remove all possible cases where a profile can be set to "" or this will continue in everything...