update getDesktopUniqueFiles to use rendererState
Now if not called explicitly with a profile it gets the rendererState's current profile.

I have been pushing to have rendererStates more consistent so they are more useful and keep better track to be used quickly in both renderer and the main process to keep everything on the same page instead of infinitely scaling and always having to pass around current profile and hoping its always on the same page. More work will be done to make rendererstates more useful and help sync the frontend and the backend to avoid problems.