Fix on previewing undefined profile on unselecting background.
Small fix for part of the problem with the default background possibly creating new profiles from nothing.

Basically now when un-selecting a background it calls to preview the new background update. If it cannot find the profile it has saved for its noBgDesktopProfile, then it hard reverts it to "default".

I don't think this covers 100% of cases, but this should essentially cover the main situations outside of manually ipc changing profile, or editing settings.json.

Theoretically this case only happens to an actual user when they create a desktopProfile, swap the default background to use it. Then delete the DesktopProfile (not yet implemented, but could manually delete the profile). In that case it would then revert back to "default" profile instead of creating a new profile with the old name.