Update DesktopGrid.css to use themes + 3 theme colors.
Added highlightOffset, highlightOversized, and highlightOffsetOversized as theme colors allowing them to be changed by the theme (if we ever allow changing of theme colors/other themes).

Converted the hard coded colors to theme colors for DesktopGrid.css

Added a colored background for the Icons overview box labels which matches their box color and makes it easier to read the box label.
Might just not make the boxes render the label unless it is non-default (moved/oversized) as if it is within its box nicely users probably don't need a box label...