textButton theme color added.
Added a basic textButton theme color. It now will be used for the default color the text in buttons.

Also updated the save button color as it is slightly different (larger).

Also removed filter button css and just converted the filter select button to normal button css (will probably come back and change a lot of buttons when UI is in the final steps).