Header + EditIcon updated to use Themes.
Updated Header + EditIcon css to use global theme colors.

Also added a header color to themes. For now it just uses the same color as bgTertiary. I wanted to go for a more like "you can't even tell a header is there" but decided against it as it might not convey the idea that the header itself is draggable. So it is just a few shades off the normal background color for most windows. (might adjust later)