Fix button text colors in EditBackground + all save buttons.
Fixed EditBackground buttons text appearing as the primary text color instead of the defined textButton color. Also fixed this for all save buttons.

Basically since button is always the primary color, doing the light/dark mode text color can be harder to read/awkward, so we have a strictly defined textButton color for it to use (which is just a white color which contrasts to be easily readable with the blue primary color).