EditBackground component embedding in BackgroundSelect.
Brief attempt to convert EditBackground into a component loaded into BackgroundSelect window instead of closing BackgroundSelect just to open EditBackground then after editing closing EditBackground just to re-open BackgroundSelect...

Basically just load it as a component in BackgroundSelect since they share a similar window shape, and avoids the closing/opening of subwindows.

Has a few problems, and could use a few tweaks for performance but a step forward for it.