Fix closing out of EditBackground not reloading background.
Fixed a problem where it would not reload the background when closing without saving a background in EditBackground.

Also fixed closing out to re-place you where you were when you opened the Edit Background, instead of re-placing the user at their selected bg.

Note: There is still a few problems with saving backgrounds in this same manner, like it does not un-preview when clicking save, and does not set the background when save & apply is selected.